+++ html/com_modules/select/default.php
@@ -19,7 +19,12 @@ if ($function)
 	HTMLHelper::_('script', 'com_modules/admin-select-modal.js', ['version' => 'auto', 'relative' => true], ['type' => 'module']);
 }
 ?>
-<h2 class="mb-3"><?php echo Text::_('COM_MODULES_TYPE_CHOOSE'); ?></h2>
+<div class="d-flex align-items-center mb-3">
+	<h2><?php echo Text::_('COM_MODULES_TYPE_CHOOSE'); ?></h2>
+	<div class="controls ml-3">
+		<input type="search" id="new-modules-list-search" class="form-control" placeholder="<?php echo Text::_('JSEARCH_FILTER'); ?>">
+	</div>
+</div>
 <div id="new-modules-list">
 	<ul class="list-group list-group-flush">
 		<?php foreach ($this->items as &$item) : ?>
@@ -29,7 +34,7 @@ if ($function)
 				<?php $name = $this->escape($item->name); ?>
 				<?php $desc = HTMLHelper::_('string.truncate', $this->escape(strip_tags($item->desc)), 200); ?>
 				<div>
-					<strong><?php echo $name; ?></strong>
+					<span class="name text-bold"><?php echo $name; ?></span>
 					<div class="text-muted"><?php echo $desc; ?></div>
 				</div>
 				<a href="<?php echo Route::_($link); ?>" class="btn btn-primary <?php echo $function ? ' select-link" data-function="' . $this->escape($function) : ''; ?>" aria-label="<?php echo Text::sprintf('COM_MODULES_SELECT_MODULE', $name); ?>">